Split kind of working except for second play it bugs out due to the print statements occurring.
