+++ src/main/java/blackjack/actions/DoubleAction.java
@@ -15,7 +15,7 @@ public class DoubleAction extends BlackJackAction {
     @Override
     public void execute(Hand playingHand, Player player, Play game) {
         player.doubleBet();
-        playingHand.addCard(game.getDeck().deal());
+        playingHand.setState(playingHand.addCard(game.getDeck().deal()));
         System.out.println(player.getName() + " doubles.");
         // TODO idk what to do, you can choose \/
         player.getPlayerManager().sendMessage(GenerateJson.generateGeneralMessage(player.toString()));
