+++ src/main/java/blackjack/player/state/BlackJack.java
@@ -1,14 +1,16 @@
 package blackjack.player.state;
 
+import blackjack.Play;
 import blackjack.actions.BlackJackAction;
 import blackjack.player.Hand;
+import blackjack.player.Player;
 
 import java.util.List;
 
 public class BlackJack implements handState {
     @Override
-    public void doRound() {
-        return;
+    public void doRound(Player player, Play game) {
+
     }
 
     @Override
