+++ src/main/java/blackjack/player/state/Bust.java
@@ -1,13 +1,15 @@
 package blackjack.player.state;
 
+import blackjack.Play;
 import blackjack.actions.BlackJackAction;
 import blackjack.player.Hand;
+import blackjack.player.Player;
 
 import java.util.List;
 
 public class Bust implements handState {
     @Override
-    public void doRound() {
+    public void doRound(Player player, Play game) {
 
     }
