+++ src/main/java/blackjack/player/state/Surrender.java
@@ -1,13 +1,15 @@
 package blackjack.player.state;
 
+import blackjack.Play;
 import blackjack.actions.BlackJackAction;
 import blackjack.player.Hand;
+import blackjack.player.Player;
 
 import java.util.List;
 
 public class Surrender implements handState {
     @Override
-    public void doRound() {
+    public void doRound(Player player, Play game) {
         return;
     }
