+++ src/main/java/blackjack/player/state/handState.java
@@ -1,13 +1,15 @@
 package blackjack.player.state;
 
+import blackjack.Play;
 import blackjack.actions.BlackJackAction;
 import blackjack.deck.Card;
 import blackjack.player.Hand;
+import blackjack.player.Player;
 
 import java.util.List;
 
 public interface handState {
-    void doRound();
+    void doRound(Player player, Play game);
 
     List<BlackJackAction> getActions(Hand hand);
 }
